feat: only request SCHEDULE_EXACT_ALARM permission if variable requires it
Context: We only need the SCHEDULE_EXACT_ALARM permission if we want to use an exact alarm to setup the background job, and we only want to do that for variables that are not activity related (e.g. STEPS), since for those we can use the Activity Recognition Transition API instead, avoiding requesting a somewhat odd permission.

References: https://outsystemsrd.atlassian.net/browse/RMET-3191