chore(android): Remove references to kapt in build.gradle
Necessary for the native library to compile with room, but not necessary for the plugin since it does not use room annotations directly.

The implementation of room library is still necessary because we use @aar for the health&fitness native library (which is useful to not come bundled with dependencies that force a certain compileSdk, because compileSdk is not controlled by the plugin), so it does not come with transitive dependencies.