fix(android): Fix Capacitor builds by removing kapt (#174)
* chore(android): Remove references to kapt in build.gradle

Necessary for the native library to compile with room, but not necessary for the plugin since it does not use room annotations directly.

The implementation of room library is still necessary because we use @aar for the health&fitness native library (which is useful to not come bundled with dependencies that force a certain compileSdk, because compileSdk is not controlled by the plugin), so it does not come with transitive dependencies.

* chore: Remove setting kapt for Capacitor apps

* chore(android): Remove unused osnotificationspermissions dependency

References:

- https://github.com/OutSystems/OSNotificationPermissionsLib-Android

* chore: prepare for version 2.4.4

References:

- https://outsystemsrd.atlassian.net/browse/RMET-4862